requirements.txt
async-timeout==5.0.1
boto3==1.37.10
botocore==1.37.10
dotenv==0.9.9
jmespath==1.0.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
s3transfer==0.11.4
six==1.17.0
urllib3==2.3.0
